ENH: Added kernels for integrands that return general objects.
It would be nice to be able to handle integrands that return arbitrary
objects that support (+) and (*), but this shouldn't come at the cost of
slowing down integrands that return doubles.

This commit adds two kernels to handle integrands that return arbitrary
objects, and adds functionality to the "run_integral" to determine
whether the integrand return value can be safely cast to float.