BLD: Allow syntax highlighting of LONG_DESCRIPTION in setup.
Previously, the code snippet in the README file (and hence in the
LONG_DESCRIPTION keyword) was just an rst literal block. Now it's
a .. code::python block.

The README file is now automatically generated from the LONG_DESCRIPTION
keyword, rather than the converse.

Because the README file has to be parsed by
Github and other programs that don't necessarily understand non-standard
extensions to rst, the substitution ".. code::python" -> "::" is made
when generating the Makefile.