BLD: Vegas import workaround.
This commit wraps the vegas import in a try statement. This is ugly but
necessary for the docs to build correctly.

The problem is that Vegas requires Cython to build correctly from
source, but readthedocs doesn't include it. Adding Cython to
scikit-monaco requirements, or even to vegas requirements, doesn't solve
the issue, which seems related to this;

https://github.com/pypa/pip/issues/1958

This commit also removes 'vegas' from the requirements.txt file, since
this is the module causing problems.